requirements.txt
# Review Master - Dependencies
# ============================

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# WhatsApp Automation
selenium>=4.8.0
webdriver-manager>=3.8.5

# Data & Utils
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
